+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+* Add formality options `'prefer_less'` and `'prefer_more'`.
+
+
 ## [1.6.0] - 2022-09-09
 ### Added
 * New language available: Ukrainian (`'uk'`). Add language code constant and
@@ -113,6 +118,7 @@ client library took over this package name. Thanks to
 ownership.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-node/compare/v1.6.0...HEAD
 [1.6.0]: https://github.com/DeepLcom/deepl-node/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/DeepLcom/deepl-node/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/DeepLcom/deepl-node/compare/v1.3.2...v1.4.0